Introduce new book adding

+++ src/main/java/com/nextit/library/dto/AnyBookDto.java
@@ -1,4 +1,4 @@
 package com.nextit.library.dto;
 
-public record AnyBookDto(long id, String name, String author, boolean borrowed) implements BookDto {
+public record AnyBookDto(int id, String name, String author, boolean borrowed) implements BookDto {
 }
