+++ src/main/java/com/nextit/library/dto/BorrowedBookDto.java
@@ -1,5 +1,5 @@
 package com.nextit.library.dto;
 
-public record BorrowedBookDto(long id, String name, String author, String borrowedTo,
+public record BorrowedBookDto(int id, String name, String author, String borrowedTo,
                               String borrowedOn) implements BookDto {
 }
